model/pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.6.1",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "torchaudio==2.8.0",
    "flash-attn",
    "tokenizers>=0.22.2",
    "transformers>=5.2.0",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.2",
    "tensorboard>=2.18.0",
    "huggingface-hub>=0.20.0",
    "opencv-python-headless>=4.13.0.92",
    "grad-cam>=1.5.5",
    "ray[serve]==2.38.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cu129" }
torchvision = { index = "pytorch-cu129" }
torchaudio = { index = "pytorch-cu129" }
flash-attn = { url = "https://github.com/Dao-AILab/flash-attention/releases/download/v2.8.3/flash_attn-2.8.3+cu12torch2.8cxx11abiTRUE-cp311-cp311-linux_x86_64.whl" }

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
